refactor(tools): review: integrate-mongodb
- reviewed mongodb integration
- moved logic to separate lib folder for global access
- minor standardizations
- fixed issue with failing web build
