+++ .env.example
@@ -1,0 +1,2 @@
+MONGODB_URI=
+MONGODB_DB_NAME=
