+++ .husky/pre-commit
@@ -1,2 +1,3 @@
 nx build web
 nx build credit-scoring
+nx build db
