--- env.example
@@ -1,2 +1,0 @@
-NEXT_PUBLIC_MONGODB_URI=
-NEXT_PUBLIC_MONGODB_DB_NAME=
