+++ libs/credit-scoring/project.json
@@ -4,20 +4,15 @@
   "sourceRoot": "libs/credit-scoring/src",
   "projectType": "library",
   "tags": [],
-  "// targets": "to see all targets run: nx show project credit-scoring --web",
   "targets": {
     "build": {
       "executor": "@nx/js:tsc",
-      "outputs": [
-        "{options.outputPath}"
-      ],
+      "outputs": ["{options.outputPath}"],
       "options": {
         "outputPath": "dist/libs/credit-scoring",
         "main": "libs/credit-scoring/src/index.ts",
         "tsConfig": "libs/credit-scoring/tsconfig.lib.json",
-        "assets": [
-          "libs/credit-scoring/*.md"
-        ]
+        "assets": ["libs/credit-scoring/*.md"]
       }
     }
   }
