+++ libs/db/README.md
@@ -1,0 +1,7 @@
+# db
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build db` to build the library.
