+++ tsconfig.base.json
@@ -17,7 +17,8 @@
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "@ibiri/credit-scoring": ["libs/credit-scoring/src/index.ts"]
+      "@ibiri/credit-scoring": ["libs/credit-scoring/src/index.ts"],
+      "@ibiri/db": ["libs/db/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
